refactor(babel-plugin-espower): extract babel-espower-visitor.js out
